+++ src/components/Game/GameContent/GameContentPlayField/EatPlayField.vue
@@ -2,16 +2,11 @@
     <div id="eat-play-field" class="d-flex flex-column">
         <div class="row justify-content-center align-items-center">
             <PlayerCard v-for="player of game.aliveWerewolfPlayers" :key="player.name" :game="game"
-                        :player="player" size="lg" class="col-3"/>
+                        :player="player" size="lg" class="col-6 col-lg-3"/>
         </div>
-        <div class="row">
-            <div class="col-12">
-                <h3 class="text-center">
-                    <VRoller :defaultChar="eatTargetText" :text="eatTargetText"/>
-                </h3>
-            </div>
-        </div>
-        <PlayerTargets :game="game" :targets="alivePlayersExceptWerewolves" :play="play" attribute="eaten" class="flex-grow-1" @playerSelected="playerSelected"/>
+        <PlayFieldActionText :game="game" :play="play" attribute="eaten" @playerSelected="playerSelected"/>
+        <PlayerTargets :game="game" :targets="alivePlayersExceptWerewolves" :play="play" attribute="eaten"
+                       class="flex-grow-1" @playerSelected="playerSelected"/>
     </div>
 </template>
 
@@ -19,10 +14,11 @@
 import PlayerCard from "../../../shared/Game/PlayerCard";
 import PlayerTargets from "../../../shared/Game/PlayerTargets/PlayerTargets";
 import Game from "../../../../classes/Game";
+import PlayFieldActionText from "../../../shared/Game/PlayField/PlayFieldActionText";
 
 export default {
     name: "EatPlayField",
-    components: { PlayerTargets, PlayerCard },
+    components: { PlayFieldActionText, PlayerTargets, PlayerCard },
     props: {
         game: {
             type: Game,
@@ -37,11 +33,6 @@ export default {
         alivePlayersExceptWerewolves() {
             return this.game.alivePlayers.filter(player => player.role.group !== "werewolves");
         },
-        eatTargetText() {
-            const playerTargeted = this.play.targets.length ? this.game.players.find(({ _id }) => _id === this.play.targets[0].player) : null;
-            const text = `${this.$tc("EatPlayField.wantToEat", this.game.aliveWerewolfPlayers.length)} `;
-            return playerTargeted ? text + playerTargeted.name : `${text} ...`;
-        },
     },
     methods: {
         playerSelected(payload) {
@@ -51,6 +42,5 @@ export default {
 };
 </script>
 
-<style scoped>
-
+<style lang="scss" scoped>
 </style>