+++ src/components/GameLobby/GameLobby.vue
@@ -8,10 +8,10 @@
                 <GameLobbyAlreadyHavePlayingGame :game="waitingGame" @cancelGame="cancelGame"/>
             </div>
             <div key="game-composition" class="p-2 d-flex flex-column h-100" v-else>
-                <div id="title" class="row justify-content-center">
+                <div class="row justify-content-center">
                     <div class="col-lg-8">
-                        <h1 class="text-center">
-                            <i class="fa fa-gamepad text-primary mr-3"/>
+                        <h1 id="title" class="text-center">
+                            <i class="fa fa-gamepad text-primary mr-2"/>
                             <span v-html="$t('GameLobby.gameComposition')"/>
                         </h1>
                         <hr class="bg-dark"/>
@@ -38,22 +38,22 @@
                     </div>
                 </div>
                 <GameLobbyComposition :game="game"/>
-                <div class="flex-grow-1">
+                <div id="game-lobby-players-container" class="flex-grow-1 container-fluid">
                     <transition mode="out-in" name="fade">
-                        <div v-if="!game.players.length" class="h-100 row justify-content-center align-items-center">
-                            <div class="row">
+                        <div v-if="!game.players.length" class="h-100 container-fluid">
+                            <div class="row h-100 justify-content-center align-items-center">
                                 <div class="col-12">
-                                    <h3 class="text-muted font-italic">
-                                        <i class="fa fa-user-plus mr-3"/>
+                                    <h3 id="no-player-text" class="text-muted text-center font-italic">
+                                        <i class="fa fa-user-plus mr-2"/>
                                         <span v-html="$t('GameLobby.addPlayerWithName')"/>
                                     </h3>
                                 </div>
                             </div>
                         </div>
-                        <transition-group v-else tag="div" name="player-item" id="players"
+                        <transition-group v-else tag="div" name="fade-list" id="players"
                                           class="row justify-content-center align-items-center h-100 p-2">
                             <PlayerCard v-for="player in game.players" :key="player.name" :game="game" :player="player"
-                                        class="player-item col-lg-2 col-xs-4" @rolePicked="rolePicked"
+                                        class="player-item col-lg-2 col-4" @rolePicked="rolePicked"
                                         @unsetRole="unsetRole" @unsetPlayer="unsetPlayer"/>
                         </transition-group>
                     </transition>
@@ -62,7 +62,7 @@
                     <hr class="bg-dark"/>
                 </div>
                 <div id="game-lobby-footer" class="row justify-content-between align-items-center">
-                    <div class="col-lg-3">
+                    <div class="col-lg-3 col-sm-6">
                         <form @submit.prevent="getGameRepartition">
                             <SubmitButton classes="btn btn-dark btn-block text-uppercase font-weight-bold"
                                           :disabled-tooltip-text="$t('GameLobby.fourPlayerRequiredToGetRandomRepartition')"
@@ -71,7 +71,7 @@
                                           :disabled="loading.createGame || !game.areThereEnoughPlayers"/>
                         </form>
                     </div>
-                    <div class="col-lg-3">
+                    <div class="col-lg-3 col-sm-6 mt-2 mt-lg-0">
                         <form @submit.prevent="createGame">
                             <SubmitButton classes="btn btn-primary btn-lg btn-block text-uppercase font-weight-bold"
                                           :text="`<i class='fa fa-play-circle mr-2'></i>${$t('GameLobby.launchParty')}`"
@@ -239,22 +239,23 @@ export default {
 };
 </script>
 
-<style>
-    .player-item {
-        transition: all 0.5s;
-    }
-    .player-item-enter, .player-item-leave-to {
-        opacity: 0;
-    }
-    .player-item-leave-active {
-        position: absolute !important;
+<style lang="scss">
+    @import "../../../node_modules/bootstrap/scss/bootstrap";
+    @import "../../assets/scss/_variables";
+
+    #title {
+        @include font-size(2rem);
     }
 
-    .flip-list-move {
-        transition: transform 1s;
+    #no-player-text {
+        @include font-size(1.5rem);
     }
 
     #player-name-input-error {
         height: 25px;
+    }
+
+    #game-lobby-players-container {
+        overflow-y: scroll;
     }
 </style>